Add flag to conditionally cleanup netmaker state

+++ scripts/nm-cleanup.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
 set -e
 
+# Arguments
+REMOVE_STATE=false
+while [ "$1" != "" ]; do
+    case $1 in
+        -a|--all)
+            REMOVE_STATE=true
+    esac
+    shift
+done
+
 # Directory containing volume data
 [ "${EUID:-$(id -u)}" -eq 0 ] \
     && NMDIR=/var/lib/netmaker \
@@ -10,14 +20,16 @@ set -e
 echo "Removing netmaker pod ..."
 podman pod exists netmaker && podman pod rm -f netmaker
 
-# Remove volumes
-echo "Removing netmaker volumes ..."
-podman volume rm -f \
-    netmaker-data \
-    netmaker-certs \
-    netmaker-mq-data \
-    netmaker-mq-logs
+if [ $REMOVE_STATE = true ]; then
+    # Remove volumes
+    echo "Removing netmaker volumes ..."
+    podman volume rm -f \
+        netmaker-data \
+        netmaker-certs \
+        netmaker-mq-data \
+        netmaker-mq-logs
 
-# Remove state directory
-echo "Removing netmaker directory ..."
-rm -rf $NMDIR
+    # Remove state directory
+    echo "Removing netmaker directory ..."
+    rm -rf $NMDIR
+fi
